+++ common/src/main/java/immersive_particles/core/registries/Renderers.java
@@ -1,7 +1,7 @@
 package immersive_particles.core.registries;
 
 import com.google.gson.JsonObject;
-import immersive_particles.core.renderers.FishRenderer;
+import immersive_particles.core.renderers.WaggleRenderer;
 import immersive_particles.core.renderers.FlappingRenderer;
 import immersive_particles.core.renderers.MeshRenderer;
 import immersive_particles.core.renderers.ParticleRenderer;
@@ -14,7 +14,7 @@ public class Renderers {
     public static final Map<String, Function<JsonObject, ParticleRenderer>> RENDERER = new HashMap<>();
 
     static {
-        register("fish", FishRenderer::new);
+        register("waggle", WaggleRenderer::new);
         register("flapping", FlappingRenderer::new);
         register("simple", MeshRenderer::new);
     }