+++ common/src/main/java/immersive_particles/core/tasks/FollowLeaderTask.java
@@ -5,6 +5,8 @@ import immersive_particles.core.ImmersiveParticle;
 import net.minecraft.util.JsonHelper;
 import org.joml.Vector3d;
 
+import java.util.Random;
+
 public class FollowLeaderTask extends Task {
     private final FollowLeaderTask.Settings settings;
 
@@ -17,10 +19,21 @@ public class FollowLeaderTask extends Task {
     public void tick() {
         ImmersiveParticle leader = particle.getLeader();
         if (leader != null) {
-            double distance = leader.getSquaredDistanceTo(particle.getPosition());
+            double distance;
+            if (particle.getTarget() != null) {
+                distance = leader.getSquaredDistanceTo(particle.getTarget());
+            } else {
+                distance = leader.getSquaredDistanceTo(particle.getPosition());
+            }
             if (distance > settings.maxDistance * settings.maxDistance) {
+                double f = (settings.maxDistance - settings.minDistance) / 2.0;
+                Random random = particle.getRandom();
                 Vector3d leaderToParticle = particle.getPosition().sub(leader.getPosition()).normalize();
-                Vector3d target = leaderToParticle.mul(settings.minDistance).add(leader.getPosition());
+                Vector3d target = leaderToParticle.mul(settings.minDistance + f).add(leader.getPosition()).add(
+                        (random.nextDouble() - 0.5) * f,
+                        (random.nextDouble() - 0.5) * f,
+                        (random.nextDouble() - 0.5) * f
+                );
                 particle.setTarget(target);
             }
         }